requirements.txt
beautifulsoup4==4.8.0
Django==2.2
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
lxml==4.4.1
pytz==2019.2
soupsieve==1.9.3
sqlparse==0.3.0
